backend/requirements.txt



fastapi
uvicorn
pillow
python-multipart

numpy==1.26.4
